professor_jon_white/quiz1_grader/requirements.txt
openai>=1.0.0
pillow>=10.0.0
pillow-heif>=0.13.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
reportlab>=4.0.0
